Move test harness tests to subdir
